fix(scraper-proxy): use lightweight utility imports
