+++ .changeset/scraper-proxy-query-cursors.md
@@ -1,5 +1,8 @@
 ---
 '@hyperlane-xyz/scraper-proxy': minor
+'@hyperlane-xyz/utils': patch
 ---
 
-Added cursor pagination to scraper message queries, with validation that cursors use a single column matching the requested sort order.
+Added cursor pagination to scraper message queries. Message cursors must contain exactly one non-null `id`; `order_by` may be omitted, in which case the cursor direction determines the ordering.
+
+Added lightweight error, set, and validation utility subpath exports.
